+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Fixed
+ - `contentFor` for buffers being strings and not found routes
 
 ## [0.1.0] - 2018-04-19
 ### Added